Fix that annoying nice-JSON-is-not-nice problem
It was because the content parameter of the copy module is not very good at dealing with
newlines, which is somewhat logical because it has to be passed through a command line.
Solution is to use really stupid templates (secret_plaintext.j2 and secret_json.j2).
That works nicely enough and it's pretty reusable, too.