azure: network load balancer is basically useless
First limitation is that you can't use a backend port both in a nat
rule and a load balancing rule. That's ok, because I can use iptables
redirect in sidethe VM. Second limitation is that you can't use the
same backend port in two different load balancing rules. That's ok
too, because I can use iptables redirect inside the VM.

The third limitation is that you can't use a frontend port for both
UDP and TCP. That just fucking sucks, because I don't know of any
way to work around it. What the fuck kind of stupid-ass arbitrary
limitation is that anyway? An UDP port is not a TCP port, there
should be no conflict there.

It seems to me that the only reasonable course of action is to
just don't use an availability set, and to have two VM's in two
Azure locations, instead of two times two VM's. This still gives
a quite reasonable uptime.

Reasoning is as follows:

Azure uptime: 99.95%; Hours in a year: 8760

Downtime for 1 machine per year  : 8760 x 0.05^1% = 4.4 hours/year
==> 99.95%
Downtime for 2 machines per year : 8760 x 0.05^2% = 13.1 minutes/year
==> 99.997%
Downtime for 3 machines per year : 8760 x 0.05^3% = 39.4 seconds/year
==> 99.9998%

I think I can probably live with 40 seconds of downtime a year. That
does suppose that my home server won't be down for more than 270
minutes a year though, not sure if I can deliver on that. Won't be
much more anyhow.